+++ src/elements/db/SegmentElementQuery.php
@@ -6,6 +6,7 @@
 
 namespace putyourlightson\campaign\elements\db;
 
+use craft\helpers\Db;
 use putyourlightson\campaign\elements\SegmentElement;
 
 use craft\elements\db\ElementQuery;
@@ -24,8 +25,30 @@ use yii\db\Connection;
  */
 class SegmentElementQuery extends ElementQuery
 {
+    // Properties
+    // =========================================================================
+
+    /**
+     * @var string|null The segment type that the resulting segments must have.
+     */
+    public $segmentType;
+
     // Protected Methods
     // =========================================================================
+
+    /**
+     * Sets the [[segmentType]] property.
+     *
+     * @param string $value The property value
+     *
+     * @return static self reference
+     */
+    public function segmentType(string $value)
+    {
+        $this->segmentType = $value;
+
+        return $this;
+    }
 
     /**
      * @inheritdoc
@@ -35,8 +58,13 @@ class SegmentElementQuery extends ElementQuery
         $this->joinElementTable('campaign_segments');
 
         $this->query->select([
+            'campaign_segments.segmentType',
             'campaign_segments.conditions',
         ]);
+
+        if ($this->segmentType) {
+            $this->subQuery->andWhere(Db::parseParam('campaign_segments.segmentType', $this->segmentType));
+        }
 
         return parent::beforePrepare();
     }